--- bin/stashcp2/.travis.yml
@@ -1,12 +1,0 @@
-
-language: python
-sudo: required
-
-notifications:
-  email:
-    recipients:
-      - taburaad@uchicago.edu
-    
-script:
-  # Run tests in Container
-  - tests/test.sh
